requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bcrypt==4.3.0
billiard==4.2.1
boto3==1.40.28
botocore==1.40.28
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.6
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
django-timezone-field==7.1
django-tinymce==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.4
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.99.9
orjson==3.11.2
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyshorteners==1.0.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stripe==12.2.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
twilio==9.6.3
Twisted==25.5.0
txaio==25.6.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
